Split fuel economy code out into files
